requirements.txt
paho-mqtt==1.6.1
protobuf==4.24.3